requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via django
billiard==4.2.1
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.6.15
    # via requests
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
cryptography==45.0.4
    # via
    #   josepy
    #   mozilla-django-oidc
distlib==0.3.9
    # via virtualenv
django==5.2.3
    # via
    #   -r requirements.in
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
    #   mozilla-django-oidc
django-redis==6.0.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
drf-yasg==1.21.10
    # via -r requirements.in
ecdsa==0.19.1
    # via python-jose
filelock==3.18.0
    # via virtualenv
identify==2.6.12
    # via pre-commit
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-yasg
iniconfig==2.1.0
    # via pytest
josepy==2.0.0
    # via mozilla-django-oidc
kombu==5.5.4
    # via celery
mozilla-django-oidc==4.0.1
    # via -r requirements.in
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   drf-yasg
    #   kombu
    #   pytest
platformdirs==4.3.8
    # via virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.2.0
    # via -r requirements.in
prompt-toolkit==3.0.51
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pygments==2.19.1
    # via pytest
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pytest==8.4.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
python-jose==3.5.0
    # via -r requirements.in
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.2
    # via
    #   drf-yasg
    #   pre-commit
redis==6.2.0
    # via
    #   -r requirements.in
    #   django-redis
requests==2.32.4
    # via mozilla-django-oidc
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via
    #   django
    #   kombu
uritemplate==4.2.0
    # via drf-yasg
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.31.2
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
